requirements.txt
requests
numpy
sentence-transformers
faiss-cpu
python-dotenv
pandas
fastapi
uvicorn
